requirements.txt
pycryptodome
pillow
numpy
matplotlib
